Improve perf of package build and remove hardcoded dependencies.
This reduced msbuild /m of packages dir from 30s to 6s on my box.  I did this by limiting the number of P2P recursions and ensuring they happen in parallel.

I've also cleaned up the cases where we hardcoded a dependency on package version and replaced with a project reference.

I've diffed the output to ensure there are no unintentional side effects.

[tfs-changeset: 1490598]