Replace use of Type.GetInterfaceMap in TypeLoader
TypeLoader uses Type.GetInterfaceMap() to find an implementation
method corresponding to a method from an interface.  But in
NET Native this method throws PlatformNotSupportedException.
As a consequence, Duplex does not work in NET Native.

This PR replaces the call to Type.GetInterfaceMap() with private
methods that find the implementation method based on the method name
and parameter types.

Fixes #153