Modify DuplexChannelFactoryTests to use APIs in public contract only
The DuplexChannelFactoryTests use APIs appearing only in the P2P APIs
that are not in the public contract. As such, these tests fail to compile when
run in the experimental CoreCLR/NETNative test environment.

It uses forms of the ctor that take an Object, which are not supported in the contract.
Replace with using teh cotr that takes an InstanceContext instead

Fixes #171