Fix exception handling in TaskMethodInvoker and work around DataContractSerializer bug
The returned task from the invoker delegate might be faulted, or the task it contains might be faulted.
We need to check both tasks for an exception. Also, there is a bug introduced in the DataContractSerializer
where it creates XmlDictionaryStrings with null IXmlDictionary's when passed the root name and namespace
as strings. We need to use a different overload or wait for the bug to be fixed. I've changed which
overload we use.