Fix dependency on System.Net.Sockets to be only for dnxcore50
.Net Native doesn't use System.Net.Sockets but the dependency was listed as a general dependency. We
depend on the 4.0.10 contract, but .Net Native only supports 4.0.0. The build system downgraded the version
to 4.0.0, which isn't actually available for dnxcore50. This meant when building the tests, System.Net.Sockets
wasn't copied across as there wasn't a matching version. We had previously worked around this by explicitly
adding that dependency to the tests. This change corrects the dependency declaration for System.Private.ServiceModel
and removes the explicit test dependencies.