Move SelfHostWcfService under tools test folder
PR #226 created a folder for tools related to tests
and moved the Bridge project into it.  This PR moves
the SelfHostWcfService into this same folder in
preparation for keeping tools and infrastructure out
of the folders meant to contain only tests.

It also silences 3 build warnings related to fields
that are not assigned to statically.