Improve BridgeClient error reporting
Adds valiation of some of the configuration
properties before they are sent to the Bridge
so failure is detected earlier.

Also captures any content body returned from
the Bridge for any non-200 response so it can
be part of the exception thrown.  This makes it
easier to see what the Bridge detected was wrong.

Fixes #227