Make Bridge service URLs configurable
This PR makes the Bridge no longer hard-code
localhost as the host for all services but instead
forms the URL based on the configuration information
passed to the Bridge POST config.

To allow this, all IResources PUT verb now accepts
a context object providing access to the Bridge
configuration.

Fixes #187