Make Bridge idle timeout configurable
This PR allows the Bridge to accept an
idle timeout as one of the properties in
the POST config name/value pairs.  And it
sets the idle timeout manager to use that
as its new idle timeout between requests.

To avoid making the Bridge aware of all its
configuration sensitive types, this PR also adds
a new EventHandler for configuration changes.
Types that care about configuration changes register
to be notified of the change.

The ConfigurationExtensions class was updated to
use this same notification mechanism.  And it was
renamed to reflect its true purpose -- AppDomainManager.

Fixes #198