Bridge.exe runs elevated and controls all Bridge commands
Bridge.exe now has an app manifest that allows it to run
as admin (UAC prompting if appropriate).

The Bridge.exe command line options have now been improved
to match  the names in TestProperties and BridgeConfiguration.
And it is possible to give Bridge.exe a .json file to initialize
multiple Bridge options.

The prior certificates used for https tests were replaced with
newer versions, and they are installed and uninstalled by the
Bridge as needed.

All firewall rule and certificate cleanup that used to happen
in other scripts is now handled by the Bridge itself. And those
prior scripts have been removed.

The BridgeController was added and a DELETE request to it will
shutdown the Bridge cleanly.