Set BridgeResourceFolder at Bridge startup
Prior to this change, the BridgeResourceFolder was specified by
BridgeClient when tests ran.  However, this was not a valid
strategy for x-platform use, otherwise both client and Bridge
would need to share a x-OS folder.

With this change, Bridge.exe is required to have a BridgeResourceFolder
when it is asked to start, and the BridgeClient no longer alters that.

Also adds a 'reset' option to Bridge.exe so that when tests complete
the Bridge can be asked to release all remote resources.  They will
be reacquired the next time Bridge requests are made.

Also move the Http DELETE handling to the ResourceController where it
really belonged.  An Http DELETE to the /resource endpoint releases all
resources.

Fixes #334