Enable OuterLoop testing for x-plat
These changes make it possible to tell the run-test.sh script
that OuterLoop tests should be included.  More work is required
to allow a CI machine to host a 2nd machine, but this PR makes
it possible to test manually.

This PR also enhances the error reporting from BridgeClient to
improve the debugging experience talking x-machine to the WCF
Bridge component.