Implement dynamic X509 client cert generation for Bridge
The bridge currently relies on static certificates for a certificate authority
cert and localhost certificate signed by the authority.
As such, cross-machine HTTP tests are dependent on a the well-known authority
certificate for server cert validation, and cross-machine TCP tests will
not work as the self-signed certs are generated for the host "localhost".

This change allows the Bridge to accept requests to generate X509 Certificates
dynamically and also serves up certificate revocation lists.

Certificates are designed to be short-lived (24h by default) but this can
be changed in the testproperties.props file along with other Bridge
configuration.

This commit addresses the following:
* X509 certificates are now generated dynamically (Fixes #324)
* Certificates are now generated via an IResource on the Bridge (Resolves #336)
* Certs generated have a Subject Alternative Name (Fixes #320)
* Support for Certificate Revocation Lists (Resolves #310)

This also addresses the following issues:
* Bridge can return raw data on a GET request (Fixes #373)
* Lay the foundation to be able to release certificates based on a
  well-known name (#335, #365 not yet fixed entirely)

This set of changes also unblocks scenarios relying on chain validation
as chain validation will now succeed - this means that tests for (say)
TCP will now work as expected, while we are waiting for new APIs to be put
in place (such as #302)

Add target reference to WcfTestBridgeCommon for package restore

Merges changes from PR #377 for allowing addressable Resource URIs