Update project.lock.json files to beta8
This updates the dependency on System.Runtime.Serialization.Xml
to 4.0.11-beta-* and updates all the project.lock.json files.

We will need to update once again if beta8 takes additional bug
fixes.  But by updating now, I found that several of our issues
that occur only in NET Native are now fixed in the beta8 packages.