Remove code dependencies on System.Net.Security for NET Native
The System.Net.Security package is not available for UWP apps,
 but WCF currently lists it as a dependency and uses it,
 regardless of platform. This causes UWP apps to fail in beta8.

The fix is to remove use of it when building for Net Native.

This PR is the first of 2. The first adds conditional compilation
 directives around code using System.Net.Security to exclude usage
 when building for NET Native.

The 2nd PR will modify the project.json and project.lock.json files
 to remove the dependency.

It is being done in 2 stages like this so that the code changes
 can be integrated into the release branch for beta8 independently.

The change to the project*.json files will require careful planning
 so that when integrated to the release branch it does not carry with
 it the new (post-beta8) dependencies currently in WCF.