Update project.json file to move System.Net.Security to dnxcore50
This is part 2 of a 2-step PR.  It updates the project.json file
to make System.Net.Security be only a dnxcore50 dependency.  It
also updates the project.lock.json to match.

Both commits of this PR are necessary in the merge to make a
version that builds in CoreCLR and NET Native in TFS.

But the first commit can be cherry-picked if necessary to move only
the code changes to the release branch.  This 2nd commit however
carries dependencies we cannot propogate into the release branch.