Validate X509CertificateClaims by SAN instead of by CN
X509CertificateClaimSet should validate based on Subject Alternative Name
instead of the Canonical Name to be consistent with Desktop .NET >= 4.6

Today, We are checking against the Subject (e.g. CN=host) on the certificate,
to determine whether or not the cert was issued for a particular server, as this
was the behaviour in <= 4.5. However, this is inflexible as the host could be
serving the same cert but bound to multiple endpoints.

In the old case, for example, foo-bar.test.com != foo-bar, and a generated cert
can only be valid for exactly one of these. In the new case, a certificate can
have a Subject with a particular CN, but multiple Subject Alt Names for which the
cert is also valid.

In the wild, most certs issued already contain SANs for the names that the
server can be, so we should be consistent with current behaviour and ditch the
legacy way of checking certs against servers.

Fixes #321