Fix commit message for "[tfs-changeset: 1538408]"
Due to infrastructure and package building reasons, we had to check in via TFS
and let the change mirror to Git. Unfortunately, the checkin message was omitted
when it was checked in via TFS - naturally the commit message in Git was blank too.
We want to make it right so that things like git blame show useful data

This reverts commit 8e06f18cf9e1af4ab353692ce79a38ab359b54ec. We will immediately
re-commit this change to fix the blank commit message that was accidentally merged
into our master branch