BUILD: Fix build to not revaluate * dependencies
Restore was regenerating lock files due to https://github.com/aspnet/dnx/issues/2771.

This caused the build to pick up "live" packages after it hit the first invalid lock file.

We are not yet ready to consume the live packages because they use generations
and many of our projects are still using "dotnet".  I need to update all projects that
 use "dotnet" to use the appropriate generation before those projects can update
their dependencies.  This is largely only a problem for our reference assembly
projects since those are the only ones that use "dotnet".  Projects that use
"dnxcore50" or "netcore50" are fine because NuGet/DNX maps that to the correct
dotnet generation.

This change fixes the problem lock files to avoid the bug.  We now expect to
no longer regenerate lock files during the build.

[tfs-changeset: 1539932]