run-test.sh no longer copies binaries from WCF bin folder
Prior to this change the run-test.sh script used to run the tests
in Linux copied all the assemblies from the WCF bin folder into the
WCF test folders.  This was harmless when the assemblies were the
same but caused issues when the bin folder contains other versions
of the assemblies.

With this change, the script no longer copies assemblies from the
WCF bin folder but relies on the test folders containing all they need.