Mark client certs installed by BridgeClient to PersistKeySet
Mark all installed certificates as PersistKeySet; this seems to be a difference when
importing to the CurrentUser certificate store vs. the LocalMachine store in
Windows. This caused issues when attemtping to use the certificates in tests and
sems to avert the issue of the tcp tests running into the following excception:

```
System.ComponentModel.Win32Exception :
The credentials supplied to the package were not recognized
```

Fixes #495