No longer copy CoreFx binaries when running Linux tests
Prior to this change, it was necessary to copy some binaries
from the CoreFx repo built on Linux when running tests.
This was necessary because the packages had incorrect binaries.

But that situation was resolved in preparation for RC1, and
it is no longer necessary to copy any binaries from CoreFx.

Fixes #442