Added a Scenario Test For XmlSerializerFormat & MessageContract.
Fix #557.

Add Scenario Tests for ServiceKnownTypeAttribute.

Fix #569.

Added a scenarios test for XmlSerializerFormat.

The test is for the case where a paramerter type contains a field never used. The test is to make sure the reflection info of the type of the unused field would be kept by Net Native toolchain.