Disable Linux tests using System.Security.Principal.Windows
These tests invoke code that causes failures in OuterLoop due to
the fact System.Security.Principal.Windows is not supported on Linux
but we still depend on it.  This causes the Windows assembly to be
used instead.

When #604 is resolved to remove this dependency we will need to
fix these tests or make them conditional on OS.