Allow Certificate Authority certificate to be exported as PEM
To test with certs in Linux, we need to import the certificate to the root cert
store as a PEM file. We don't currently allow certificates to be exportable
easily (they get returned as JSON key-pairs for parsing at the client side).

With this change we allow the cert to be retrieved from the Bridge in PEM format
so we can script an import as necessary.