Update rd.xml to allow instantiation of open generic FaultException
The NET Native toolchain has been improved to allow the instantiation
of open generics via a new entry in the rd.xml file.  This PR adds that
new line.

The rd.xml already had a similar line that permitted this
for generic types where the T was a class. That line has been left
unmodified to mitigate risk if a new WCF package is used with an
older toolchain.

An app was built containing this new syntax and compiled against the
older toolchain to demonstrate the new syntax would not break it.

Fixes #664