Make Bridge honor max idle timeout set from command line
Prior to this change, the Bridge initialized its max idle
timeout to a fixed value and did not adjust for a value being
passed on the commandline at startup.

It was necessary to fix this issue so that a Bridge launched
on a lab machine can be given a maximum idle timeout.  This
permits the Bridge to close gracefully regardless how a test
run completes.