Add directive to rd.xml for MessageHeader.CreateHeader()
When XmlSerializerFormatAttribute is marked on a OperationContract, .Net Native tool assums that the operation wants to serialize objects using XmlSerializer and so pre-generates XmlSerializer for all relevant types. However, MessageHeader.CreateHeader() internally used DataContractSerializer to serialize the header object, even if the operation is marked with XmlSerializerFormatAttribute.

The fix is adding directives in the rd.xml to tell .Net Native tool to generate DataContractSerializer for the 'value' parameter of the MessageHeader.CreateHeader() method.

I also added tests covering the case of using MessageHeader.CreateHeader together with XmlSerializerFormatAttribute.

Fix #740