Adds new /require option to Bridge.exe to allow optional start
The new /require option will ensure the Bridge is running.
If it is already running, it just returns.  But if it is not
running it starts the Bridge.exe in a new process and waits for
it to respond to pings.

This new option is not yet used but will be in automated runs
on Windows machines that require the Bridge is running before
a test can execute.