Allow Bridge to know whether it was manually started
To enable automation of the Bridge in OuterLoop test
runs in the lab, it is necessary for the Bridge itself
to know whether it was manually started by the developer
or automatically started by a test.

The StartBridge.cmd script now propagates that information
to the Bridge.  And a new -stopIfAutoStart switch to Bridge.exe
will stop the Bridge only if it was not manually started.

The old KeepBridgeRunning environment variable is no longer used.