Remove BridgeAutoStart and let scripts decide how to close Bridge
The BridgeAutoStart property has been removed because it was too
confusing.  Instead, 'Bridge.exe /require' will choose whether or
not to start the Bridge and return an exit code with its decision.

EnsureBridgeRunning.cmd sets an environment variable based on this
exit code so that our normal build.cmd scripts can choose to stop
the Bridge only if the script started it.  The rule is "If you
start the Bridge, you are responsible for its shutdown. But if you
find it already running, you are allowed to reset it (release its
resources) but not to shut it down."

This retains the ability to use StartBridge to obtain a running
Bridge that tests will not automatically teardown.