Simplify and branchify the netci file
This does some basic refactoring and simplification of the CI definition, as well as moving WCF to the branch based model (see below)

Move wcf to branch model

This moves wcf to the branch-based CI model.  In this model, instead of having the CI generate jobs out of dotnet/wcf's master branch for all branches (using all sorts of odd naming suffixes of course), the master CI lists tell the CI which branches it should look at for CI config info.  It passes the branch name to the netci file, which then can use it to tell the SCM to pull code from specific branches or set branch specific PR triggers.