Allow MsBuild properties at build time to flow to the Bridge.
This PR makes it possible for Bridge configuration properties
expressed at build time to be built into the test binaries.
And it also allows Build.exe to accept the MSBuild style of
property arguments.

This enables 2 important scenarios for our multi-machine effort.

1. Doing "build /p:BridgeHost=xyz" will generate the appropriate
TestProperties at compile time so that the generated test binaries
carry this information. Then if they are uploaded to Helix to
execute, they already carry the information necessary to communicate
with a Bridge on a different machine.  There is no need to set
environment variables on the execution machine for this.

2. By allowing Bridge.exe to parse the MSBuild style property
arguments, it allows our build scripts to flow Bridge configuration
information to the Bridge startup/ping scripts so they agree with
what the tests expect.

Prior to this PR, it would have been necessary to set environment
variables on both the client and execution machines to have a
shared Bridge configuration.