Map ERROR_INVALID_HANDLE to EndpointNotFoundException
On Linux, HttpClient returns an HResult with ERROR_INVALID_HANDLE
attempting to access an endpoint it cannot find. As a result, we
throw CommunicationException rather than EndpointNotFoundException.

This is inconsistent with Windows behavior and breaks a scenario
test for this case.

Until CoreFx provides a common way to map HResults to a predictable
set of values across platforms, we must map this specific HResult
manually in our exception conversion helpers.

Fixes #419