Fix test failure when BasicAuth tests were first to run
The Https BasicAuthentication tests failed were they were the first
scenario tests run after a fresh start of the Bridge.  But they succeeded
when other scenario tests ran first.

Root cause was that the BasicAuthResource was not explicitly ensuring
the SSL port certificate was installed.  It passed if another scenario
test did that first.

The fix was to duplicate what the other Https test resources do and
explicitly ensure the SSL port certificate is installed.

Fixes #788