Update crypto dependencies
System.Security.Cryptography.Algorithms and
System.Security.Cryptography.X509Certificates have updated their
versions to 4.1 and now target both NETStandard1.3 and
NETStandard1.4.

Update references to these packages and update any project that is
restoring for DNXCore50 but packaging as NETStandard to restore for
NETStandard as well so that it gets the correct references.