Fix restore errors after upgrading packages
The System.Private.ServiceModel windows build was depending on package
versions that didn't include support for the targeted platform.  Upgrade
these to satisfy guardrails.

Some test projects were depending on non-existent package versions
(major.minor.build invalid) which have been updated to valid versions.

xunit.extensibility.execution was bringing in an old version of
Linq.Expressions that doesn't support DNXCore50, add a dependency to
any project consuming this to update that dependency to latest.

We removed the lifting from the lineup so tests should reference the
latest packages in order to ensure an implementation is available.