Fix NullReferenceException in SendRequestAsync during async streaming
Problem: HttpChannelFactory.SendRequestAsync keeps the HttpRequestMessage
in a member field and uses it during async processing.  But a Fault or Abort
will null the HttpRequestMessage field.  If this happens while the
SendRequestAsync method is executing, it leads to NullReferenceException
during error handling.

Solution: keep a local variable copy of the HttpRequestMessage for use in
error handling.  Even if the original is disposed or the field set to null,
error handling only requires access to the RequestUri property.

Fixes #931