Adds scenario test for AV during Http async stream read
Issue #931 was discovered by stress tests, where an aborted
Http request doing an asynchronous read would throw a
NullReferenceException.

The cause is not a race condition but instead incorrect use of
member fields in exception handling.  This PR adds a scenario test
that exactly duplicates the stress test and causes NullRefException.

This test will be merged only after #935 that contains the fix to
issue #931.  The changes are being kept separate to allow selective
cherry-picking into the RC2 branch.