Add support for [ConditionalFact] based on TestProperties
We need a way to optionally run tests that are not normally
run by default.  This includes tests that have special requirements
for setup or environment.

We also need such tests to show as "skipped" rather than being
silently stripped from the test results, as happens with the
xunit -notrait option.

This PR introduces the framework to compose these kinds of test
conditions using TestProperties and adds some of the known conditions.
Adding [ConditionalFact] to individual tests will be done in separate PR's.