Fix TFMs for System.Private.ServiceModel deps
The TFM for these dependencies is incorrect, causing the S.P.SM to be
missed when not targeting netstandardapp1.5 (eg for a portable app).

Fix this TFM to match the implementation TFM.

I've filed a buildtools issue https://github.com/dotnet/buildtools/issues/617
to help eliminate the need for this completely.