Fix Digest auth test to work when username and password are available
Remove [ActiveIssue] from this test and add to exiting [ConditionalFact]
that it should run only when username and password has been provided in
TestProperties. Default is that they are not provided, in which case
this test is reported as skipped.

This test has been verified to run and pass using a command like:
  build.cmd /p:TestUserName=xxx /p:TestPassword=yyy ...
where xxx and yyy are known good domain credentials.  It has also
been proven to run and fail when given bad domain credentials.

Fixes #979