Update [ConditionalFact] certificate detectors to install and check
Changes the certificate [ConditionalFact] detectors to attempt to
install the certificates but then also to verify that they exist
in the store.  This permits scenarios where the install can fail
but a certificate is already in the store.

This change also renames and refactors BridgeClientCertificateManager
to remove all unused Bridge code and to simplify for only the certificate
operations currently required.