Remove certificate matching by name
The ConditionalFact detector for certificates was using a fallback
strategy to locate certificates based on IssuedBy and SubjectName
when it could not locate by thumbprint.  But this allowed matches
on left-over certificates that would not work.

With this change, we match only on thumbprint or report that there
is no certificate installed and let the ConditionalFact skip the
certificate tests.

Fixes #1172