Update pre-release tags to RC4
The long term plan is to move our packages versions up and the
prerelease tag back down to something like -beta, but for now we just
need to pick something that isn't RC3 as that's what
the release/1.0.0 branches are using.

Doing this gives us time to do the right thing without having version
clashes on myget.

[tfs-changeset: 1606983]