Fix package baseline
Package baselining has been broken since the first package update after
I added it.  The problem was that the package that contains the corefx
baseline was missing from the update list so the update script updated
it to be out of sync with the CoreFxExpectedPrerelease property. I've
added the baseline package to the list of packages coming from CoreFx.

This was silently being dropped because we don't import the target if it
doesn't exist.  I've updated the import condition to be less permissive.
We just need it to not import before the restore so constraining it to
pkgproj will do that, and fail the build if this issue should occur in
the future.

Port of https://github.com/dotnet/wcf/commit/c5118f33c380761e2c738fdeb06e3ec411ab15d2