Update PRService to be robust when fetching new branches
The PRService looks up a list of valid branches before allowing a checkout to a
branch on the query string. However, when a new branch is created, the PRService
does not do a fetch before doing the branch name check, which means the PRService
is not aware of the new branch and fails the check.

This PR
1. Causes the fetch to happen first
2. Consolidates the fetch/cleanup logic into CleanupBranches

Fixes #1287