Remove ActiveIssue(1223) to re-enable tests that failed in CI
We used ActiveIssue(1123) to disable some failing CI tests
when we were first bringing up the new test infrastructure.

This PR replaces ActiveIssue(1123) with a new ConditionalFact.
This allows these tests to continue to run on machines known
to be properly configured, and it allows CI and lab runs to
tell the tests when the test client has been configured properly.