Fix MessageInterceptor test
The MessageInterceptor scenario test was attempting to work around
the product issue where custom channels and channel factories were
not being opened or closed properly.  This workaround made the tests
work, but was an incorrect approach.  On the desktop, these same tests
failed for issue #1241.

This commit removes the workaround from the MessageInterceptor tests
so that they exactly mirror the corresponding sample code found
[here](https://msdn.microsoft.com/library/ms751495(v=vs.100).aspx).

Combined with the product fix, these tests now pass.  Moreover,
when the prior workarounds are combined with the product fix, we now
fail in NET Core the same way we fail against the full framework.
So now we have behavioral parity with the full framework for both
the positive (no workaround) and negative (with workaround) cases.