Change test certificate acquisition to REST endpoint
The certificate acquisition endpoint currently used requires a WCF client.
This causes some complications when our scripts need to obtain certificates,
as we need to spin up a WCF client capable of understanding a WCF response.

This change moves certificate acquisition to come from a REST endpoint instead
so we can (for example) use cURL to acquire the certificate instead of using
a complicated workaround like the TestRootCertificateInstaller