Add new [WcfFact] and [WcfTheory] attributes to generalize ConditionalFact
This PR adds 2 new attributes and discoverers to take the place of
ConditionalFact and ConditionalTheory.  The new discoverers have
been generalized to reflect onto the test method for any of the new
"skip" attributes (IssueAttribute and ConditionAttribute).

It also adds enums for OSID and FrameworkID to identify the
current OS and Framework, and it adds optional parameters to
the new IssueAttribute to allow it to be applied only to specific
OSes or Frameworks.

For design discussion on this see issue #1034