Add better checks for detecting detached HEAD in PRService (#1406)
Add better checks for detecting detached HEAD in PRService

Without the check for a deteached HEAD state, we would attempt to delete some
bogus branches. Now we're a little better about it and so we should see the
PRService fail less often

P.S. the algorithm would probably not pass if asked in an interview. Please do
     not copy for such purposes