Move to xunit 2.2.0-prerelease to get UWP testing to work
Preparing to get UWP testing working in Helix discovered
WcfFactAttribute and WcfTheoryAttribute were not working
properly.  One reason was due to mismatched xunit
versions between the UWP runners and our Infrastructure.Common.
Moving to the same xunit version allows that part to work.

The second reason was that our custom Fact and Theory attributes
and their discoverers were not permitted to do Reflection.  The fix
was to add a custom rd.xml that allows that.  With these 2 changes
WcfFact and WcfTheory work in UWP in Helix.