Fix UWP issue #1450 and re-enable tests disabled for it.
Issue #1450 was caused in UWP because the MemberData attribute
was hiding specific Encoding types from the reducer, resulting in
MissingMetadataException when xunit was doing test discovery.

The fix is to add back a custom rd.xml file to the project containing
the TestData class to grant Reflection permissions for types affected by this.

Fixes #1450