Enhance OSID detection when running under NET Native
RuntimeInformation.OSDescription is hard-coded to return a generic
"Microsoft Windows" description when compiled for NET Native. This
masks from our OSID detection which version of Windows is running.

Prior to this fix, we returned "AnyWindows" to indicate we did not
know, but this caused some IssueAttributes to skip Win7 tests.

The fix is to refine the OSID detection slightly when the hard-coded
"Microsoft Windows" description is returned.

Fixes #1459