Omit empty ETW "Begin" and "End" messages if no events to report
The new logic to Console.WriteLine ETW events for tests that fail
is writing out extraneous empty "Begin" and "End" blocks when run
on UWP.  This is because UWP doesn't currently support ETW, and it
has no events to report.

The fix is to skip the entire Console.WriteLine block of code if
there are no ETW events to report.

Fixes #1465