Enable WS* layers of WCF
This enabling will happening in several stages.
This first stage is porting the missing types into the GitHub
repo, code formatting, and commenting out code that is not needed to allow
SecurityBindingElement.CreateUserNameOverTransportBindingElement
to work.

Subsequent PR's will add additional tests and expand the WS*
functionality a piece at a time.