Fix issue where CommunicationObject.OnOpen was not called
Converting the synchronous CommunicationObject open/close paths
introduced a bug that prevented the OnOpen() method to be called
when the synchronous Open() method was used.

The fix is to be aware of whether the caller invoked the synchronous
Open or the asynchronous BeginOpen and invoke the appropriate
synchronous or asynchronous override.

Fixes #1544