Reference dependencies of common test projects
WCF uses 3 common test projects:
- System.Private.ServiceModel\tests\Common\Infrastructure\Infrastructure.Common.csproj
- System.Private.ServiceModel\tests\Common\Scenarios\ScenarioTests.Common.csproj
- System.Private.ServiceModel\tests\Common\Unit\UnitTests.Common.csproj

Since these are referenced through MSBuild-based project references and
we're using dotnet.exe to restore, the packages don't flow across the
project reference.  As a result the tests were running with an older
version of binaries for things that were referenced by the common
project but not the test project.  Fix this by ensuring all dependencies
referenced by the common projects are also part of the test-runtime.